Add minRows parser method

+++ module/ldbc-generator/src/main/scala/ldbc/generator/model/Table.scala
@@ -27,7 +27,7 @@ object Table:
     case class InsertMethod(value: "NO" | "FIRST" | "LAST") extends Options
     case class KeyBlockSize(value: Int)                     extends Options
     case class MaxRows(value: Long)                          extends Options
-    case class MinRows(value: Int)                          extends Options
+    case class MinRows(value: Long)                          extends Options
     case class PackKeys(value: "0" | "1" | "DEFAULT")       extends Options
     case class Password(value: String)                      extends Options
     case class RowFormat(value: "DEFAULT" | "DYNAMIC" | "FIXED" | "COMPRESSED" | "REDUNDANT" | "COMPACT")
